dev authentication and first page and home page
Not finished yet
