+++ NinerFit/Authentication/views.py
@@ -1,0 +1,45 @@
+from django.shortcuts import render, redirect
+from django.contrib import auth
+from django.contrib.auth.models import User
+
+#login function
+def login(request):
+    if request.method == 'POST':
+        username = request.POST['username']
+        password = request.POST['password']
+        user = auth.authenticate(request, username=username, password=password)
+        if user is not None:
+            auth.login(request, user)
+            return redirect('chatbot')
+        else:
+            error_message = 'Invalid username or password'
+            return render(request, 'login.html', {'error_message': error_message})
+    else:
+        return render(request, 'login.html')
+
+#Register function
+def register(request):
+    if request.method == 'POST':
+        username = request.POST['username']
+        email = request.POST['email']
+        password1 = request.POST['password1']
+        password2 = request.POST['password2']
+
+        if password1 == password2:
+            try:
+                user = User.objects.create_user(username, email, password1)
+                user.save()
+                auth.login(request, user)
+                return redirect('chatbot')
+            except:
+                error_message = 'Error creating account'
+                return render(request, 'register.html', {'error_message': error_message})
+        else:
+            error_message = 'Password dont match'
+            return render(request, 'register.html', {'error_message': error_message})
+    return render(request, 'register.html')
+
+#Logout Function
+def logout(request):
+    auth.logout(request)
+    return redirect('login')
