+++ NinerFit/ChatBot/views.py
@@ -5,6 +5,8 @@ from django.contrib import auth
 from django.contrib.auth.models import User
 from .models import Chat
 from django.utils import timezone
+from django.contrib.auth.decorators import login_required
+
 
 #set open AI API
 your_api = '' #insert own api key here
@@ -25,6 +27,7 @@ def ask_openai(message):
     return answer
 
 # Create your views here.
+@login_required
 def chatbot(request):
     chats = Chat.objects.filter(user=request.user)
     if request.method == 'POST':
@@ -34,45 +37,3 @@ def chatbot(request):
         chat.save()
         return JsonResponse({'message': message, 'response': response})
     return render(request, 'chatbot.html', {'chats': chats})
-
-#login function
-def login(request):
-    if request.method == 'POST':
-        username = request.POST['username']
-        password = request.POST['password']
-        user = auth.authenticate(request, username=username, password=password)
-        if user is not None:
-            auth.login(request, user)
-            return redirect('chatbot')
-        else:
-            error_message = 'Invalid username or password'
-            return render(request, 'login.html', {'error_message': error_message})
-    else:
-        return render(request, 'login.html')
-
-#Register function
-def register(request):
-    if request.method == 'POST':
-        username = request.POST['username']
-        email = request.POST['email']
-        password1 = request.POST['password1']
-        password2 = request.POST['password2']
-
-        if password1 == password2:
-            try:
-                user = User.objects.create_user(username, email, password1)
-                user.save()
-                auth.login(request, user)
-                return redirect('chatbot')
-            except:
-                error_message = 'Error creating account'
-                return render(request, 'register.html', {'error_message': error_message})
-        else:
-            error_message = 'Password dont match'
-            return render(request, 'register.html', {'error_message': error_message})
-    return render(request, 'register.html')
-
-#Logout Function
-def logout(request):
-    auth.logout(request)
-    return redirect('login')
