+++ NinerFit/main/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+class MainConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'main'  # Ensure this is lowercase and matches the folder name
