+++ NinerFit/main/templates/home.html
@@ -1,0 +1,3 @@
+<h1>Home Page</h1>
+<p>Welcome, {{ user.username }}!</p>
+<a href="{% url 'chatbot' %}">Go to Chatbot</a> | <a href="{% url 'logout' %}">Logout</a>
