+++ NinerFit/main/templates/start.html
@@ -1,0 +1,2 @@
+<h1>Welcome to Our App</h1>
+<p>Please <a href="{% url 'login' %}">Login</a> or <a href="{% url 'register' %}">Register</a> to continue.</p>
